pyproject.toml
[project]
name = "williamzujkowski-blog"
version = "1.0.0"
description = "William Zujkowski's personal blog and portfolio - automation scripts and content management"
requires-python = ">=3.11"
authors = [
    {name = "William Zujkowski", email = "william@zujkowski.com"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Core web scraping and HTTP
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "certifi>=2023.0.0",

    # Data formats and parsing
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",

    # Markdown processing
    "markdown>=3.5.0",

    # Browser automation
    "playwright>=1.40.0",

    # Utilities
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",

    # Image processing
    "pillow>=10.0.0",

    # Text analysis
    "textstat>=0.7.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",

    # Code quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",

    # Type stubs
    "types-pyyaml>=6.0",
    "types-requests>=2.31",
    "types-python-dateutil>=2.8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | node_modules
  | _site
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "frontmatter.*",
    "playwright.*",
    "textstat.*",
]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
